Work on cookies and login state - feature: work on cookies and logging in state to keep user logged in - refactor: replace Signin.js to root src directory
