+++ frontend/package.json
@@ -33,6 +33,7 @@
     "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-chartist": "^0.14.4",
+    "react-cookie": "^4.1.1",
     "react-copy-to-clipboard": "^5.0.3",
     "react-datetime": "^3.0.4",
     "react-dom": "^16.13.1",
