+++ frontend/src/index.js
@@ -15,7 +15,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { HashRouter } from "react-router-dom";
+import { BrowserRouter } from 'react-router-dom/cjs/react-router-dom.min';
 
 // core styles
 import "./scss/volt.scss";
@@ -27,9 +27,9 @@ import HomePage from "./pages/HomePage";
 import ScrollToTop from "./components/ScrollToTop";
 
 ReactDOM.render(
-  <HashRouter>
+  <BrowserRouter>
     <ScrollToTop />
     <HomePage />
-  </HashRouter>,
+  </BrowserRouter>,
   document.getElementById("root")
 );
